requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.3
